Bail out the configuration of Sentry Logs if Odoo has it already.
Our current setup (buildout) may setup the logging before Odoo gets to parse
the args.  This is, the initialization script calls the patch_logging which,
in turn, initialize the logging.  At this point Odoo has not processed the
`--log*` arguments yet.

The idea is to embed the sentry setup within Odoo itself.  This means that
mailgate needs to call the sentry setup itself.  That's why we provide the
force argument.