Rework the registry.
Remove usage of deprecated RegistryManager when using Odoo 10.  Most of the
functions of the RegistryManager are now classmethods of Registry.

Split the registry implementation in several modules.  This is just to let me
know where I am.

Since the backport of Odoo 10's shell into xoeuf, we can use the shell like::

  $ xoeuf shell -d somedb

  >>> self
  res.users(1, )

After this commit, to access any DB we keep the 'from xoeuf.pool import db'
idiom but it's been totally changed.  Now, when you salt the shell you only
get a 'self' with ``res.users()`` almost like as with the `-d` argument.  This
is::

  $ xoeuf shell

  >>> from xoeuf.pool import db
  >>> db.salt_shell()

  >>> self
  res.users()

This still a work progress.  There are probably many things that can be
simplified or removed given the limit scope of the xoeuf registry.

I'm merging this into develop to allow for developers to test it in the wild.