Fix unstable shell commands.
Without putting the newly created class in the module globals it may
disappear: the Command's metaclass does not keep a reference until latter when
the commands are listed, so the newly created class may be garbage-collected.

Without this commit any shell command (ie. 'xoeuf python') had some
probability to be missing and we'd get the help message.