Make the transaction manager to wrap a whole environment.
The transaction manager used to wrap a DB cursor directly.  With the move to
Odoo 10, cursors are no longer passed as arguments.  The old API is no more.
The only reason to access a cursor is to execute plain SQL.

When checking the context, we do have to create our own Environment and access
the cursor to avoid a recursive call.