Add a line profiler tool.
WARNING: You should never install `line_profiler` in production and let a
profiler activated.

Within a development environment this tools allows to profile specific
functions to find performance bottlenecks in your code.

For instance, I'm using it to detect why the invoices take so long to
validate.  This is bulk of the contents inside `profiler.rst`.