Allow Property to be *inherited*.
Fixes the issue shown in commit 33ebe4a37711ddb7a69853c08132cbeef2f5de19.

When a module `_inherits` other it makes a copy of the fields, but with the
``inherited`` property set, and also make the field a related one.  The
low-level `_write()` method of Model finds those fields and issue the proper
`write()` on related fields.

This is done in:

```
            v = {}
            for fname in updend:
                field = self._fields[fname]
                if field.inherited and field.related[0] == inherit_field:
                    v[fname] = vals[fname]
                    unknown_fields.discard(fname)
            if v:
                self.pool[table].write(cr, user, nids, v, context)
```

for Odoo 8 and 9.

Odoo 10 changes that a bit:

```
            parent_vals = {}
            for name in updend:
                field = self._fields[name]
                if field.inherited and field.related[0] == parent_field:
                    parent_vals[name] = vals[name]
                    unknown_fields.discard(name)

            if parent_vals:
                self.env[parent_model].browse(parent_ids).write(parent_vals)
```

In all cases a `write` is issue on the model/table from which we
are *inheriting*.

So we only need to let honor the attributes ``inherited`` and ``related`` for
Property.