Remove CLI command mailgate.
We're now using an run of the mill inbox and have the fetchmail addon check
the inbox each minute.

The main reason behind this change is to avoid loosing emails if, while
executing Odoo code, an exception happens.  Having the emails safely stored in
the Inbox allows to, first, fix the bug/exception and re-fetch the emails.