Allow TimeField descriptor to take full `time` values.
When passing a string, it was limited to 'HH:MM' values; which is inconsistent
with allowing to pass `time` instances.

There's still one inconsistency: tzinfo.  Which we note.

When printing the time value in a TimeRange, the *best* format is chosen.  The
invariant is that ``eval(repr(time_range)) == time_range``.