Create valid TimeRange in tests.
If the randomly generated times `t1` and `t2` were closer than the precision
of the format `fmt` would allow, we could create an invalid TimeRange.

For ``t1 = time(0, 0, 1)``, ``t2 = time(0, 0, 0, 1)`` and ``fmt = '%H:%M'``,
the test would try to create ``TimeRange(time(0, 0, 0, 1), '00:00')``.

The solution is not to test the formatting feature of TimeRange in the same
test and create a very similar one, for that purpose.