Refactor the fields.Enumeration to avoid signals.
Those signals were being executed for all models.  But the amount of models
using fields.Enumeration is rather small.  So this could represent a
performance hit that can be avoided.

The new implementation uses the `setup_full` method for fields to setup insert
an *adapter* class in the MRO the models that actually use an Enumeration
field.