Refactor signals tests to improve readability.
The 'self.mocks' was hiding too much and hinder comprehension of each test.
We MUST trade clarity over re-usability in tests.

The names of the receivers, although correct for normal naming conventions --
they state that they do, not what they're connected to; for testing the
signals they're supposed to be connected with, it's best to name them
according to their connections.  After all, they do nothing.