Proof that setting values in generalized Enumerations may fail.
The tests show that this only manifest in char-based enumerations.

The problem seems to be that fields __set__ does:

            write_value = self.convert_to_write(self.convert_to_record(value, record), record)
            record.write({self.name: write_value})

So our 'write' method gets the value already converted to str, and tries to
find that in the members.

In the Integer based there's no problem because the value to write is also the
integer (not the name).