Fix Enumeration when setting inside a computed field.
This fixes the tests written in commits
8ce5b227c91a17303fbfd79134924267d772bb7a and
74ae957541f5c273870ee09bed19db3638777353.

Since the issue is indeed that 'write' can be called after calling
convert_to_write, we simply return the proper member in 'convert_to_write' and
let our 'write' method actually convert it from member to name.

We also remove the 'convert_to_record' and instead make the member a part of
the cache.

'convert_to_column' is necessary because if the column has a default, Odoo
executes may take that value to create the SQL column.  Again only needed for
the fields.Char based Enumerations.