Fix 'not in', 'not like' and 'not ilike' for mapped/filtered filters.
With a non-mapped filter the previous approach 'not . in' only worked for
simple attributes.  But for x2many fields, Odoo regards `('children.age', 'not
in', value)` as 'any children has an age not in value'; our previous approach
was 'not (any children has an age in value)' which is equivalent to 'all
children has an age not in value'.

Therefore 'not in' is not the same as 'not . in'.