Fix putting Enumeration fields in an AbstractModel.
We should find the EnumerationAdapter in the MRO and in the current model's
bases only.  Otherwise we inject the EnumerationAdapter twice when using the
Enumeration field in a AbstractModel: in the class of the abstract model and
the class of the model.  This raises a TypeError: Cannot create a consistent
method resolution order (MRO) for bases ...

This was a bug in the current implementation of Enumeration field (we need to
enforce testing with AbstractModel is all fields).