Fix issues in the selection field compute/set methods.
When the selection field is created the field hasn't yet been *attached* to
the model (setup_full); so the `self` we have inside the method
get_selection_field won't have the Enumclass: to attach a field to model Odoo
must make a copy of the field with ``field.new()`` and the call
``setup_full()``.

So we must resolve the enumclass inside the _compute/_set method of the
selection field.