Force recomputation of computed fields depending on a Property.
Now instead of waiting for dependant fields to get the values (via higher
level field access); we actively trigger the recomputation of fields which
depend on a Property.

This means setting a Property may impact the DB for those fields and that an
afterwards call to `read()` will get the right value.