Implement _get_filter_ast of domains without relying on DomainTree.
Remove both DomainTree._get_filter_ast() and deprecates DomainTree.walk().
Deprecation follows from the fact the walk may have a undefined order because
DomainTree.children is a set.

Two DomainTree on differently ordered domains could yield the walk:

   >>> DomainTree(Domain([('x', '!=', False), ('x', '<' , "equis")]).second_normal_form)
   (('x', '<', 'equis') '&' ('x', '!=', False))

   >>> DomainTree(Domain([('x', '<', 'equis'), ('x', '!=', False)]).second_normal_form)
   (('x', '<', 'equis') '&' ('x', '!=', False))

That changes the semantics of the domain with regard to short-circuit
operations.  I don't know if the simplify algorithm is affected.

Fixes #8.