Add the timezone where the time range to be represented.
The computing of the time range from a Datetime field is done with the value
in UTC and therefore doesn't correspond to the date-time by the user set from
the web interface.

For example, set from the user interface '2020-07-25 08:00' from the zone
'America / Havana' (-4) the result is that in UTC it is at noon.